refactor: move SERVER_ERROR constant from authorization.ts to api.ts
